requirements.txt
openai
python-dotenv
pyttsx3
speechrecognition
pyaudio
